editorium/app/requirements1.txt
#-i https://download.pytorch.org/whl/cu118
torch==2.6.0
torchvision
controlnet_aux
OmegaConf==2.3.0
xformers
ctransformers==0.2.27
auto-gptq==0.7.1
#diffusers>=0.30.3
#accelerate>=0.34.2
accelerate
# git+https://github.com/huggingface/diffusers.git@main#egg=diffusers
git+https://github.com/rodjjo/diffusers.git@main#egg=diffusers
transformers>=4.48.2
numpy==1.26.4
sentencepiece>=0.2.0
SwissArmyTransformer>=0.4.12
gradio>=4.44.0
imageio>=2.35.1
imageio-ffmpeg>=0.5.1
openai>=1.45.0
moviepy>=1.0.3
pillow==10.4.0
flask>=2.1.1
flask-cors>=5.0.0
spandrel==0.4.0
# opencv-contrib-python==4.10.0.84
opencv-python-headless==4.8.0.74
sk-video==1.1.10
safetensors==0.4.4
peft==0.13.0
decord==0.6.0
bitsandbytes==0.45.1
prodigyopt==1.0
requests==2.32.3
huggingface-hub>=0.27.0
# flash-attn==2.6.3
einops==0.8.0
basicsr@git+https://github.com/rodjjo/BasicSR@master#egg=basicsr
gfpgan
realesrgan==0.3.0
rembg==2.0.59
# mpi4py==4.0.1
# deepspeed==0.15.2
marshmallow==3.22.0
#basicsr@git+https://github.com/XPixelGroup/BasicSR.git@033cd6896d898fdd3dcda32e3102a792efa1b8f4
scikit-image==0.24.0
blendmodes==2024.1.1
websocket-server==0.6.4
gguf>=0.10.0
timm==1.0.15
ftfy==6.3.1
